Module Gg.Box2

2D axis-aligned boxes.

type t = box2

The type for 2D boxes (rectangles).

val dim : int

dim is the dimension of the boxes of type box2.

type v = v2

The type for 2D vectors.

type p = p2

The type for 2D points.

type size = size2

The type for 2D sizes.

type m = m2

The type for matrices representing linear transformations of 2D space.

Constructors, accessors and constants

val v : p2 -> size2 -> box2

v o size is a box whose origin is o and size is size.

val v_mid : p2 -> size2 -> t

v_mid mid size is a box whose mid point is mid and size is size.

val empty : box2

empty is the empty box.

val o : box2 -> p2

o b is the origin of b.

Raises Invalid_argument on empty

val ox : box2 -> float

ox b is V2.x (o b).

val oy : box2 -> float

oy b is V2.y (o b).

val size : box2 -> size

size b is the size of b.

Raises Invalid_argument on empty

val w : box2 -> float

w b is Size2.w (size b).

val h : box2 -> float

h b is Size2.h (size b).

val zero : box2

zero is a box whose origin and size is zero.

val unit : box2

unit is the unit box which extends from zero to one in all dimensions.

val of_pts : p2 -> p2 -> box2

of_pts p p' is the smallest box whose space contains p and p'.

val add_pt : box2 -> p2 -> box2

add_pt b p is the smallest box whose space contains b and p.

Functions

val min : box2 -> p2

min b is the smallest point of b (its origin).

Raises Invalid_argument on empty

val minx : box2 -> float

minx b is V2.x (min b).

val miny : box2 -> float

miny b is V2.y (min b).

val max : box2 -> p2

max b is the greatest point of b (its size added to the origin).

Raises Invalid_argument on empty

val maxx : box2 -> float

maxx b is V2.x (max b).

val maxy : box2 -> float

maxy b is V2.y (max b).

val mid : box2 -> p2

mid b is the mid point between min and max.

Raises Invalid_argument on empty

val midx : box2 -> float

midx b is V2.x (mid b).

val midy : box2 -> float

midy b is V2.y (mid b).

val bl_pt : box2 -> p2

bl b is the bottom-left corner of b.

Raises Invalid_argument on empty

val bm_pt : box2 -> p2

bm_pt b is the bottom-mid point of b.

Raises Invalid_argument on empty

val br_pt : box2 -> p2

br_pt b is the bottom-right corner of b.

Raises Invalid_argument on empty

val ml_pt : box2 -> p2

ml_pt b is the mid-left corner of b.

Raises Invalid_argument on empty

val mm_pt : box2 -> p2

ml_pt b is {!mid} b.

Raises Invalid_argument on empty

val mr_pt : box2 -> p2

mr_pt b is the mid-right point of b.

Raises Invalid_argument on empty

val tl_pt : box2 -> p2

tl_pt b is the top-left corner of b.

Raises Invalid_argument on empty

val tm_pt : box2 -> p2

tm_pt b is the top-middle corner of b.

Raises Invalid_argument on empty

val tr_pt : box2 -> p2

tr_pt b is the top-right corner of b.

Raises Invalid_argument on empty

val area : box2 -> float

area b is the surface area of b.

val inter : box2 -> box2 -> box2

inter b b' is a box whose space is the intersection of S(b) and S(b').

val union : box2 -> box2 -> box2

union b b' is the smallest box whose space contains S(b) and S(b').

val inset : v2 -> box2 -> box2

inset d b is b whose edges are inset in each dimension according to amounts in d. Negative values in d outset. If the size in dimension i becomes negative it is clamped to 0 and the ith coordinate of the mid point of b is used for the ith coordinate of the resulting box's origin. Returns empty on empty.

val outset : v2 -> box2 -> box2

outset d b is inset (V2.neg d) b.

val round : box2 -> box2

round b is the smallest box containing b with integer valued corners. Returns empty on empty.

val move : v2 -> box2 -> box2

move d b is b translated by d. Returns empty on empty.

val ltr : m2 -> box2 -> box2

ltr m b is the smallest box containing the corners of b transformed by m. Returns empty on empty.

val tr : m3 -> box2 -> box2

tr m b is the smallest box containing the corners of b transformed by m in homogenous 2D space. Returns empty on empty.

val map_f : (float -> float) -> box2 -> box2

map_f f b is the box whose origin and size are those of b with their components mapped by f. Returns empty on empty.

Predicates and comparisons

val is_empty : box2 -> bool

is_empty b is true iff b is empty.

val is_pt : box2 -> bool

is_pt b is true iff b is not empty and its size is equal to 0 in every dimension.

val is_seg : box2 -> bool

is_seg b is true iff b is not empty and its size is equal to 0 in exactly one dimension.

val isects : box2 -> box2 -> bool

isects b b' is not (is_empty (inter b b')).

val subset : box2 -> box2 -> bool

subset b b' is true iff S(b) is included in S(b').

val mem : p2 -> box2 -> bool

mem p b is true iff p is in S(b).

val equal : box2 -> box2 -> bool

equal b b' is b = b'.

val equal_f : (float -> float -> bool) -> box2 -> box2 -> bool

equal_f eq b b' tests b and b' like equal but uses eq to test floating point values.

val compare : box2 -> box2 -> int

compare u v is Stdlib.compare u v.

val compare_f : (float -> float -> int) -> box2 -> box2 -> int

compare_f cmp b b' compares b and b' like compare but uses cmp to compare floating point values.

Formatters

val pp : Stdlib.Format.formatter -> box2 -> unit

pp ppf b prints a textual representation of b on ppf.

val pp_f : (Stdlib.Format.formatter -> float -> unit) -> Stdlib.Format.formatter -> box2 -> unit

pp_f pp_fl ppf b prints b like pp but uses pp_fl to print floating point values.