Module type Gg.Size

Implemented by all size types.

type t

The type for sizes.

val dim : int

dim is the dimension of sizes of type t.

Constructors, accessors and constants

val zero : t

zero is the zero size, zero extent in each dimension.

val unit : t

unit is the unit size, one extent in each dimension.