Module Gg.Size1

In 1D space, width is the extent along the x-axis.

type t = float

The type for 1D sizes.

val dim : int

dim is the dimension of sizes of type size1

Constructors, accessors and constants

val v : float -> size1

v w is a size of width w.

val w : size1 -> float

w s is the width of s.

val zero : size1

zero is the zero size, zero width.

val unit : size1

unit is the unit size, one width.