Module Gg.Size3

In 3D space, width is the extent along the x-axis, height the extent along the y-axis and depth the extent along the z-axis.

type t = size3

The type for 3D sizes.

val dim : int

dim is the dimension of sizes of type size3.

Constructors, accessors and constants

val v : float -> float -> float -> size3

v w h d is a size of width w, height h and depth d.

val w : size3 -> float

w s is the width of s.

val h : size3 -> float

h s is the height of s.

val d : size3 -> float

d s is the depth of s.

val zero : size3

zero is the zero size, zero width, height and depth.

val unit : size3

unit is the unit size, one width, height and depth.